// src/app/users/user.type.ts
export interface UserProfile {
  id: string;
  created_at: string;
  updated_at: string;
  is_deleted: boolean;
  role: string;
  email: string;
  phone: string | null;
  is_verified: boolean;
  is_active: boolean;
  is_admin: boolean;
  otp_id: number;
  onesignal_id: string | null;
  is_verification_pending: boolean | null;
  avatar: string;
  capture_avatar: string | null;
  name: string;
  date_of_birth: string;
  gender: string;
  nid_photos: string[];
  driving_license_photos: string[];
  vehicle_type: string | null;
  vehicle_brand: string | null;
  vehicle_model: string | null;
  vehicle_plate_number: string | null;
  vehicle_registration_photos: string[];
  vehicle_photos: string[];
  trip_given_count: number;
  trip_received_count: number;
  stripe_account_id: string | null;
  is_stripe_connected: boolean;
  rating: number;
  rating_count: number;
  location_type: string;
  location_lat: number | null;
  location_lng: number | null;
  location_address: string | null;
  is_online: boolean;
  last_online_at: string;
}